Extracted `struct Instrumentation{Point,PointBuilder,Adder,Builder}` to `crate::point`.  This will be further refactored.
